+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2023-07-18
+
+### Added
+- Extra debug information (writting command queue + formatting)
+
+### Changed
+- Now SendCommand method is static
+
 ## [1.1.0] - 2023-07-07
 
 ### Added